Change form

+++ base/forms.py
@@ -4,7 +4,7 @@ from datetime import datetime
 
 class weightForm(forms.Form):
     date = forms.DateField(label="Date", widget=forms.SelectDateWidget, initial=datetime.now())
-    weight = forms.IntegerField(label="Weight", min_value=0, max_value=500)
+    weight = forms.FloatField(label="Weight", min_value=0, max_value=500)
 class editDayForm(forms.Form):
     date = forms.DateField(label="Date you wish to edit:")
     new_cal = forms.IntegerField(label="New calories:", min_value=0, max_value=20000)
